+++ libs/flexIPC/fleXdUDS/examples/ChangeLog.md
@@ -2,7 +2,15 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
+## [0.0.1] - 2018-06-12
+### Added
+- None
 
+### Changed
+- Updated server and client examples [Martin Strenger]
+
+### Removed
+- None
 
 ## [0.0.1] - 2018-03-10
 ### Added